# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "instadroid",
    "summary": "Atom feeds of an Instagram Following feed scraped from a real Android app.",
    "description": "When the server runs with `FEED_TOKEN`, every path except `/health` needs that token: as `Authorization: Bearer <token>`, as the password of HTTP basic auth, or as `?token=`. `/media/<file>` URLs written into the feeds carry a per-file `sig` instead.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/control/scrape-now": {
      "post": {
        "summary": "Ask for a run now",
        "description": "Ask the poll loop to start a run within about 30 seconds instead of waiting out its sleep.\nRefused while locked, and within RUN_NOW_MIN_MINUTES of the last run finishing.",
        "operationId": "scrape_now_control_scrape_now_post",
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ControlState"
                }
              }
            }
          },
          "409": {
            "description": "The manual lock is in place."
          },
          "429": {
            "description": "Too soon."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ControlState": {
        "properties": {
          "locked": {
            "type": "boolean",
            "title": "Locked"
          },
          "scrape_now": {
            "type": "boolean",
            "title": "Scrape Now"
          },
          "hold": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hold"
          }
        },
        "type": "object",
        "required": [
          "locked",
          "scrape_now"
        ],
        "title": "ControlState"
      },
...
    }
  }
}